+++ test/e2e/tests/tracing.go
@@ -8,19 +8,14 @@
 package tests
 
 import (
-	"context"
 	"fmt"
-	"testing"
-	"time"
-
+	"github.com/envoyproxy/gateway/internal/utils/naming"
+	"github.com/envoyproxy/gateway/test/utils/tracing"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/wait"
 	httputils "sigs.k8s.io/gateway-api/conformance/utils/http"
 	"sigs.k8s.io/gateway-api/conformance/utils/kubernetes"
 	"sigs.k8s.io/gateway-api/conformance/utils/suite"
-	"sigs.k8s.io/gateway-api/conformance/utils/tlog"
-
-	"github.com/envoyproxy/gateway/internal/utils/naming"
+	"testing"
 )
 
 func init() {
@@ -55,22 +50,7 @@ var OpenTelemetryTracingTest = suite.ConformanceTest{
 				"provider":     "otel",
 				"service.name": naming.ServiceName(gwNN),
 			}
-			// let's wait for the log to be sent to stdout
-			if err := wait.PollUntilContextTimeout(context.TODO(), time.Second, time.Minute, true,
-				func(ctx context.Context) (bool, error) {
-					count, err := QueryTraceFromTempo(t, suite.Client, tags)
-					if err != nil {
-						tlog.Logf(t, "failed to get trace count from tempo: %v", err)
-						return false, nil
-					}
-
-					if count > 0 {
-						return true, nil
-					}
-					return false, nil
-				}); err != nil {
-				t.Errorf("failed to get trace from tempo: %v", err)
-			}
+			tracing.ExpectedTraceCount(t, suite, gwAddr, expectedResponse, tags)
 		})
 	},
 }
@@ -105,39 +85,7 @@ var ZipkinTracingTest = suite.ConformanceTest{
 				// should make them kept consistent
 				"service.name": fmt.Sprintf("%s/%s", gwNN.Namespace, gwNN.Name),
 			}
-			if err := wait.PollUntilContextTimeout(context.TODO(), time.Second, time.Minute, true,
-				func(ctx context.Context) (bool, error) {
-					preCount, err := QueryTraceFromTempo(t, suite.Client, tags)
-					if err != nil {
-						tlog.Logf(t, "failed to get trace count from tempo: %v", err)
-						return false, nil
-					}
-
-					httputils.MakeRequestAndExpectEventuallyConsistentResponse(t, suite.RoundTripper, suite.TimeoutConfig, gwAddr, expectedResponse)
-
-					// looks like we need almost 15 seconds to get the trace from Tempo?
-					err = wait.PollUntilContextTimeout(context.TODO(), time.Second, 15*time.Second, true, func(ctx context.Context) (done bool, err error) {
-						curCount, err := QueryTraceFromTempo(t, suite.Client, tags)
-						if err != nil {
-							tlog.Logf(t, "failed to get curCount count from tempo: %v", err)
-							return false, nil
-						}
-
-						if curCount > preCount {
-							return true, nil
-						}
-
-						return false, nil
-					})
-					if err != nil {
-						tlog.Logf(t, "failed to get current count from tempo: %v", err)
-						return false, nil
-					}
-
-					return true, nil
-				}); err != nil {
-				t.Errorf("failed to get trace from tempo: %v", err)
-			}
+			tracing.ExpectedTraceCount(t, suite, gwAddr, expectedResponse, tags)
 		})
 	},
 }
@@ -170,39 +118,7 @@ var DatadogTracingTest = suite.ConformanceTest{
 				"provider":     "datadog",
 				"service.name": fmt.Sprintf("%s.%s", gwNN.Name, gwNN.Namespace),
 			}
-			if err := wait.PollUntilContextTimeout(context.TODO(), time.Second, time.Minute, true,
-				func(ctx context.Context) (bool, error) {
-					preCount, err := QueryTraceFromTempo(t, suite.Client, tags)
-					if err != nil {
-						tlog.Logf(t, "failed to get trace count from tempo: %v", err)
-						return false, nil
-					}
-
-					httputils.MakeRequestAndExpectEventuallyConsistentResponse(t, suite.RoundTripper, suite.TimeoutConfig, gwAddr, expectedResponse)
-
-					// looks like we need almost 15 seconds to get the trace from Tempo?
-					err = wait.PollUntilContextTimeout(context.TODO(), time.Second, 60*time.Second, true, func(ctx context.Context) (done bool, err error) {
-						curCount, err := QueryTraceFromTempo(t, suite.Client, tags)
-						if err != nil {
-							tlog.Logf(t, "failed to get curCount count from tempo: %v", err)
-							return false, nil
-						}
-
-						if curCount > preCount {
-							return true, nil
-						}
-
-						return false, nil
-					})
-					if err != nil {
-						tlog.Logf(t, "failed to get current count from tempo: %v", err)
-						return false, nil
-					}
-
-					return true, nil
-				}); err != nil {
-				t.Errorf("failed to get trace from tempo: %v", err)
-			}
+			tracing.ExpectedTraceCount(t, suite, gwAddr, expectedResponse, tags)
 		})
 	},
 }
